Create Motor and Transistor Fan
This code first shows a transistor working to control the power to the motor. As explained in my data sheet on transistors, they often work as switches in order to manipulate the power. As you can see in the first line of code, the transistor is turned on and off using pin 3. Next, the user is reminded that the motor speeds lies within integers 0 and 255 and the user is able to fix the motor to a desired input integer speed.